Moved CSV generation into a separate file.
